Cflores-addings3

+++ server/utils/cleanupFiles.js
@@ -3,15 +3,20 @@ const fs = require('fs');
 function cleanupFiles(filePaths) {
   return Promise.all(filePaths.map(filePath => {
     return new Promise((resolve, reject) => {
-      fs.unlink(filePath, (err) => {
-        if (err) {
-          console.error('Error deleting file:', err);
-          reject(err);
-        } else {
-          console.log('File deleted successfully:', filePath);
-          resolve();
-        }
-      });
+      if (fs.existsSync(filePath)) {
+        fs.unlink(filePath, (err) => {
+          if (err) {
+            console.error('Error deleting file:', err);
+            reject(err);
+          } else {
+            console.log('File deleted successfully:', filePath);
+            resolve();
+          }
+        });
+      } else {
+        console.log(`File not found: ${filePath}`);
+        resolve();  // Resolve the promise if the file is not found
+      }
     });
   }));
 }
